requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.21
google-api-python-client==2.100.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
requests==2.31.0

